move beacons out of redundant beacons module
