+++ app/views/beacons/_form.html.erb
