+++ app/views/beacons/edit.html.erb
