+++ app/views/beacons/index.html.erb
