+++ app/views/beacons/new.html.erb
