+++ app/views/beacons/show.html.erb
